setup haptic
