+++ Sofa/Component/Compat/src/SofaHaptics/ForceFeedback.h
@@ -19,69 +19,14 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaHaptics/initSofaHaptics.h>
+#pragma once
 
-#include <sofa/helper/system/PluginManager.h>
+#include <sofa/component/haptic/ForceFeedback.h>
 
-namespace sofa
+// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptic/ForceFeedback.h")
+
+namespace sofa::component::controller
 {
+    using ForceFeedback = sofa::component::haptic::ForceFeedback;
 
-namespace component
-{
-
-
-void initSofaHaptics()
-{
-    static bool first = true;
-    if (first)
-    {
-        // msg_deprecated("SofaHaptics") << "SofaHaptics is deprecated. It will be removed at v23.06. Use Sofa.Component.Haptic instead.";
-
-        sofa::helper::system::PluginManager::getInstance().loadPlugin("Sofa.Component.Haptic");
-
-        first = false;
-    }
-}
-
-extern "C" {
-SOFA_SOFAHAPTICS_API void initExternalModule();
-SOFA_SOFAHAPTICS_API const char* getModuleName();
-SOFA_SOFAHAPTICS_API const char* getModuleVersion();
-SOFA_SOFAHAPTICS_API const char* getModuleLicense();
-SOFA_SOFAHAPTICS_API const char* getModuleDescription();
-SOFA_SOFAHAPTICS_API const char* getModuleComponentList();
-}
-
-void initExternalModule()
-{
-    initSofaHaptics();
-}
-
-const char* getModuleName()
-{
-    return "SofaHaptics";
-}
-
-const char* getModuleVersion()
-{
-    return "1.0";
-}
-
-const char* getModuleLicense()
-{
-    return "LGPL";
-}
-
-const char* getModuleDescription()
-{
-    return "This module contains the base infrastructure for haptics rendering in Sofa.";
-}
-
-const char* getModuleComponentList()
-{
-    return "NullForceFeedback LCPForceFeedback";
-}
-
-} // namespace component
-
-} // namespace sofa
+} // namespace sofa::component::controller
