+++ Sofa/Component/Compat/src/SofaHaptics/NullForceFeedbackT.h
@@ -19,43 +19,14 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_CONTROLLER_nullptrFORCEFEEDBACKT_H
-#define SOFA_COMPONENT_CONTROLLER_nullptrFORCEFEEDBACKT_H
-#include <SofaHaptics/MechanicalStateForceFeedback.h>
-#include <SofaHaptics/config.h>
-namespace sofa
+#pragma once
+
+#include <sofa/component/haptic/NullForceFeedbackT.h>
+
+// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptic/NullForceFeedbackT.h")
+
+namespace sofa::component::controller
 {
+    using NullForceFeedbackT = sofa::component::haptic::NullForceFeedbackT;
 
-namespace component
-{
-
-namespace controller
-{
-
-/// @brief Null force feedback for haptic feedback device
-template<class TDataTypes>
-class SOFA_SOFAHAPTICS_API NullForceFeedbackT : public sofa::component::controller::MechanicalStateForceFeedback<TDataTypes>
-{
-    typedef TDataTypes DataTypes;
-    typedef typename DataTypes::VecCoord VecCoord;
-    typedef typename DataTypes::VecDeriv VecDeriv;
-
-public:
-    SOFA_CLASS(SOFA_TEMPLATE(NullForceFeedbackT,TDataTypes),sofa::component::controller::MechanicalStateForceFeedback<TDataTypes>);
-    void init() override {this->ForceFeedback::init();}
-
-    void computeForce(SReal, SReal, SReal, SReal, SReal, SReal, SReal, SReal& fx, SReal& fy, SReal& fz) override
-    {
-        fx = fy = fz = 0.0;
-    }
-    void computeForce(const  VecCoord &,  VecDeriv &) override {}
-    void computeWrench(const sofa::defaulttype::SolidTypes<SReal>::Transform &, const sofa::defaulttype::SolidTypes<SReal>::SpatialVector &, sofa::defaulttype::SolidTypes<SReal>::SpatialVector &W_tool_world ) override {W_tool_world.clear();}
-};
-
-} // namespace controller
-
-} // namespace component
-
-} // namespace sofa
-
-#endif
+} // namespace sofa::component::controller
