+++ Sofa/Component/Haptic/Sofa.Component.HapticConfig.cmake.in
@@ -1,0 +1,12 @@
+# CMake package configuration file for @PROJECT_NAME@
+
+@PACKAGE_GUARD@
+@PACKAGE_INIT@
+
+find_package(Sofa.Simulation.Core QUIET REQUIRED)
+find_package(Sofa.Component.Constraint.Lagrangian.Solver QUIET REQUIRED)
+
+if(NOT TARGET @PROJECT_NAME@)
+    include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
+endif()
+check_required_components(@PROJECT_NAME@)
