+++ Sofa/Component/Haptic/src/sofa/component/haptic/ForceFeedback.h
@@ -19,49 +19,50 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_CONTROLLER_MECHANICALSTATEFORCEFEEDBACK_H
-#define SOFA_COMPONENT_CONTROLLER_MECHANICALSTATEFORCEFEEDBACK_H
+#pragma once
 
-#include <SofaHaptics/config.h>
-#include <SofaHaptics/ForceFeedback.h>
+#include <sofa/component/haptic/config.h>
+
 #include <sofa/simulation/fwd.h>
+#include <sofa/core/behavior/BaseController.h>
+#include <sofa/defaulttype/SolidTypes.h>
+#include <sofa/defaulttype/RigidTypes.h>
 
-namespace sofa
+namespace sofa::component::haptic
 {
 
-namespace component
-{
-
-namespace controller
-{
-
-template<class TDataTypes>
-class SOFA_SOFAHAPTICS_API MechanicalStateForceFeedback : public sofa::component::controller::ForceFeedback
+/// Base class implementing forcefeedback as a force field
+class SOFA_COMPONENT_HAPTIC_API ForceFeedback : public virtual core::behavior::BaseController
 {
 
 public:
-    SOFA_CLASS(SOFA_TEMPLATE(MechanicalStateForceFeedback,TDataTypes),sofa::component::controller::ForceFeedback);
+    SOFA_ABSTRACT_CLASS(ForceFeedback,core::behavior::BaseController);
+    Data<bool> d_activate; ///< boolean to activate or deactivate the forcefeedback
+    Data<int> d_indice; ///< Tool indice in the OmniDriver
 
-    typedef TDataTypes DataTypes;
-    typedef typename DataTypes::VecCoord VecCoord;
-    typedef typename DataTypes::VecDeriv VecDeriv;
+    simulation::Node *context;
 
-public:
-    virtual void computeForce(const  VecCoord& state,  VecDeriv& forces) = 0;
+    void init() override;
 
-    void init() override {context = sofa::simulation::node::getNodeFrom(getContext());}
-    void computeForce(SReal x, SReal y, SReal z, SReal u, SReal v, SReal w, SReal q, SReal& fx, SReal& fy, SReal& fz) override = 0;
-    void computeWrench(const sofa::defaulttype::SolidTypes<SReal>::Transform &, const sofa::defaulttype::SolidTypes<SReal>::SpatialVector &, sofa::defaulttype::SolidTypes<SReal>::SpatialVector & ) override = 0;
-    void setReferencePosition(sofa::defaulttype::SolidTypes<SReal>::Transform& /*referencePosition*/) override {}
+    virtual void computeForce(SReal x, SReal y, SReal z,
+                              SReal u, SReal v, SReal w,
+                              SReal q, SReal& fx, SReal& fy, SReal& fz) = 0;
+
+    virtual void computeWrench(const sofa::defaulttype::SolidTypes<SReal>::Transform &,
+                               const sofa::defaulttype::SolidTypes<SReal>::SpatialVector &,
+                               sofa::defaulttype::SolidTypes<SReal>::SpatialVector & )=0;
+
+    virtual void setReferencePosition(sofa::defaulttype::SolidTypes<SReal>::Transform& referencePosition);
+    virtual bool isEnabled();
+
+    /// Abstract method to lock or unlock the force feedback computation. To be implemented by child class if needed
+    virtual void setLock(bool value)
+    {
+        SOFA_UNUSED(value);
+    }
 
 protected:
-    MechanicalStateForceFeedback(void) {}
+    ForceFeedback();
 };
 
-} // namespace controller
-
-} // namespace component
-
-} // namespace sofa
-
-#endif
+} // namespace sofa::component::haptic
