+++ Sofa/Component/Haptic/src/sofa/component/haptic/LCPForceFeedback.cpp
@@ -21,17 +21,11 @@
 ******************************************************************************/
 #define SOFA_COMPONENT_CONTROLLER_LCPFORCEFEEDBACK_CPP
 
-#include <SofaHaptics/LCPForceFeedback.inl>
+#include <sofa/component/haptic/LCPForceFeedback.inl>
 #include <sofa/core/ObjectFactory.h>
 #include <sofa/defaulttype/RigidTypes.h>
 
-namespace sofa
-{
-
-namespace component
-{
-
-namespace controller
+namespace sofa::component::haptic
 {
 
 using sofa::defaulttype::Rigid3Types;
@@ -76,12 +70,8 @@ int lCPForceFeedbackClass = sofa::core::RegisterObject("LCP force feedback for t
         .add< LCPForceFeedback<defaulttype::Vec1Types> >()
         .add< LCPForceFeedback<defaulttype::Rigid3Types> >();
 
-template class SOFA_SOFAHAPTICS_API LCPForceFeedback<defaulttype::Vec1Types>;
-template class SOFA_SOFAHAPTICS_API LCPForceFeedback<defaulttype::Rigid3Types>;
+template class SOFA_COMPONENT_HAPTIC_API LCPForceFeedback<defaulttype::Vec1Types>;
+template class SOFA_COMPONENT_HAPTIC_API LCPForceFeedback<defaulttype::Rigid3Types>;
 
 
-} // namespace controller
-
-} // namespace component
-
-} // namespace sofa
+} // namespace sofa::component::haptic
