+++ Sofa/Component/Haptic/src/sofa/component/haptic/config.h.in
@@ -19,69 +19,20 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaHaptics/initSofaHaptics.h>
+#pragma once
 
-#include <sofa/helper/system/PluginManager.h>
+#include <sofa/config.h>
+#include <sofa/config/sharedlibrary_defines.h>
 
-namespace sofa
+#ifdef SOFA_BUILD_SOFA_COMPONENT_HAPTIC
+#  define SOFA_TARGET @PROJECT_NAME@
+#  define SOFA_COMPONENT_HAPTIC_API SOFA_EXPORT_DYNAMIC_LIBRARY
+#else
+#  define SOFA_COMPONENT_HAPTIC_API SOFA_IMPORT_DYNAMIC_LIBRARY
+#endif
+
+namespace sofa::component::haptic
 {
-
-namespace component
-{
-
-
-void initSofaHaptics()
-{
-    static bool first = true;
-    if (first)
-    {
-        // msg_deprecated("SofaHaptics") << "SofaHaptics is deprecated. It will be removed at v23.06. Use Sofa.Component.Haptic instead.";
-
-        sofa::helper::system::PluginManager::getInstance().loadPlugin("Sofa.Component.Haptic");
-
-        first = false;
-    }
-}
-
-extern "C" {
-SOFA_SOFAHAPTICS_API void initExternalModule();
-SOFA_SOFAHAPTICS_API const char* getModuleName();
-SOFA_SOFAHAPTICS_API const char* getModuleVersion();
-SOFA_SOFAHAPTICS_API const char* getModuleLicense();
-SOFA_SOFAHAPTICS_API const char* getModuleDescription();
-SOFA_SOFAHAPTICS_API const char* getModuleComponentList();
-}
-
-void initExternalModule()
-{
-    initSofaHaptics();
-}
-
-const char* getModuleName()
-{
-    return "SofaHaptics";
-}
-
-const char* getModuleVersion()
-{
-    return "1.0";
-}
-
-const char* getModuleLicense()
-{
-    return "LGPL";
-}
-
-const char* getModuleDescription()
-{
-    return "This module contains the base infrastructure for haptics rendering in Sofa.";
-}
-
-const char* getModuleComponentList()
-{
-    return "NullForceFeedback LCPForceFeedback";
-}
-
-} // namespace component
-
-} // namespace sofa
+	constexpr const char* MODULE_NAME = "@PROJECT_NAME@";
+	constexpr const char* MODULE_VERSION = "@PROJECT_VERSION@";
+} // namespace sofa::component::haptic
