+++ Sofa/Component/Haptic/tests/scenes/ToolvsFloorCollision_test.scn
@@ -1,8 +1,8 @@
 <?xml version="1.0" ?>
 <Node name="root" dt="0.05" showBoundingTree="0" gravity="0 -1 0">
-    <RequiredPlugin name="Sofa.Component" />
+    <RequiredPlugin name="Sofa.Component"/>
     <RequiredPlugin name="Sofa.GL.Component" />
-    <RequiredPlugin name="SofaHaptics"/>
+    <RequiredPlugin name="Sofa.Component.Haptic"/>
     
     <DefaultPipeline name="pipeline" depth="6" verbose="0"/>
     <BruteForceBroadPhase/>
