+++ modules/SofaHaptics/CMakeLists.txt
@@ -4,37 +4,23 @@ project(SofaHaptics LANGUAGES CXX)
 set(HEADER_FILES
     src/SofaHaptics/config.h
     src/SofaHaptics/initSofaHaptics.h
-    src/SofaHaptics/ForceFeedback.h
-    src/SofaHaptics/LCPForceFeedback.h
-    src/SofaHaptics/LCPForceFeedback.inl
-    src/SofaHaptics/MechanicalStateForceFeedback.h
-    src/SofaHaptics/NullForceFeedback.h
-    src/SofaHaptics/NullForceFeedbackT.h
     )
 
 set(SOURCE_FILES
     src/SofaHaptics/initSofaHaptics.cpp
-    src/SofaHaptics/ForceFeedback.cpp
-    src/SofaHaptics/LCPForceFeedback.cpp
-    src/SofaHaptics/NullForceFeedback.cpp
-    src/SofaHaptics/NullForceFeedbackT.cpp
     )
 
 set(EXTRA_FILES
     README.md
     )
 
-find_package(Sofa.Simulation.Core REQUIRED)
-sofa_find_package(Sofa.Component.Constraint.Lagrangian.Solver REQUIRED)
-
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES} ${EXTRA_FILES})
-target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Simulation.Core)
-target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Constraint.Lagrangian.Solver)
 
-if (NOT "${CMAKE_CXX_COMPILER_ID}" STREQUAL "MSVC")
-    # Silence attribute warnings (for example, ignored already defined external template)
-    target_compile_options(${PROJECT_NAME} PRIVATE -Wno-attributes)
-endif()
+message(WARNING "${PROJECT_NAME} module is deprecated. It will be removed at v23.06. Use Sofa.Component.Haptic instead.")
+
+# forward to the new sofang module
+sofa_find_package(Sofa.Component.Haptic REQUIRED)
+target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Haptic)
 
 sofa_create_package_with_targets(
     PACKAGE_NAME ${PROJECT_NAME}
@@ -44,10 +30,3 @@ sofa_create_package_with_targets(
     INCLUDE_INSTALL_DIR "${PROJECT_NAME}"
     RELOCATABLE "plugins"
     )
-    
-# Tests
-cmake_dependent_option(SOFAHAPTICS_BUILD_TESTS "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
-if(SOFAHAPTICS_BUILD_TESTS)
-    enable_testing()
-    add_subdirectory(SofaHaptics_test)
-endif()
